Create factories dir in products app
