+++ apps/products/factories/__init__.py
@@ -1,0 +1,1 @@
+from .factories import *  # noqa
